backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
telethon==1.32.1
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
